POC: Map and objects should be in JSON
Proof of concept: Map and objects are in JSON-file
Not complete, but there is no need to continue because JSON-functionality will be rewritten in future. Still dummy version, where json-data is moved to json-file, but creation of objects still is done inside code and you have to know what json-has.